Fixes: #8098
